Refactor code to update thread message count at regular intervals
This commit updates the code in the backend/db.js file to modify the interval for updating the thread message count. The previous interval was set to 1 minute, but it has been changed to 5 minutes. Additionally, a setTimeout function is added to update the count after 300 milliseconds.

These changes ensure that the thread message count is updated at regular intervals, improving the accuracy of the count in the database.